web-services/python-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
requests==2.31.0
prometheus-client==0.17.1